chore: moved tab 3 to the user icon & other things
