+++ src/styles/style.css
@@ -2,18 +2,18 @@
 
 :root {
   --color-0: #0f1115;
-  --color-1: #1A1D24;
+  --color-1: #1a1d24;
   --color-1b: #1a1d24ea;
-  --color-2: #22252E;
+  --color-2: #22252e;
   --color-3: #2c2e3a;
   --color-4: #32333f;
   --color-important: #ff456d;
   /* #ffc4c9 */
   --color-dark-opacity: rgba(0, 0, 0, 0.6);
-  --color-warning: #FF5454;
-  --color-alert: #E5A639;
-  --color-success: #45AD67;
-  --color-info: #3FA0D0;
+  --color-warning: #ff5454;
+  --color-alert: #e5a639;
+  --color-success: #45ad67;
+  --color-info: #3fa0d0;
 
   --font-1: #f1f2f6;
   --font-2: #bec0cc;
@@ -224,11 +224,11 @@ img {
 }
 
 .auto-update-page .content .buttons button:hover {
-  filter: brightness(0.80);
+  filter: brightness(0.8);
 }
 
 .auto-update-page .content .buttons button:active {
-  filter: brightness(0.70);
+  filter: brightness(0.7);
 }
 
 /* MAIN SECTION */
@@ -313,7 +313,7 @@ main section .scrollers {
 main.search {
   display: flex;
   flex-direction: row;
-  gap: 10px
+  gap: 10px;
 }
 
 main .filters-container {
@@ -456,12 +456,12 @@ main button#search-clear {
 
 main button#search-submit:hover,
 main button#search-clear:hover {
-  filter: brightness(0.90);
+  filter: brightness(0.9);
 }
 
 main button#search-submit:active,
 main button#search-clear:active {
-  filter: brightness(0.80);
+  filter: brightness(0.8);
 }
 
 main .entries-container {
@@ -602,5 +602,139 @@ main .entries-container .anime-entry {
 }
 
 .settings-page button#exit:active {
-  filter: brightness(0.90);
+  filter: brightness(0.9);
+}
+
+.user-page .user-info {
+  gap: 20px;
+  display: flex;
+  flex-direction: row;
+}
+
+.user-page .user-info .user-name p {
+  color: var(--font-2);
+  font-weight: 700;
+  font-size: 2rem;
+  line-height: 28px;
+  margin-top: 4.3rem;
+  font-family: 'DM Sans', sans-serif;
+}
+
+.user-page .user-info .user-logout {
+  background-color: var(--color-important);
+  color: var(--font-1);
+  font-size: 1.5rem;
+  margin-top: 11px;
+  font-weight: 700;
+  width: 100%;
+  height: 25%;
+  border-radius: 0.5rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+}
+
+.user-page .user-info .user-avatar img {
+  width: 150px;
+  height: auto;
+  border-radius: 0.5rem;
+}
+
+.user-page .user-info .user-stats {
+  height: 5rem;
+  display: grid;
+  gap: 10px;
+  margin-top: 24px;
+  padding: 0.25rem;
+  margin-left: auto;
+  text-align: center;
+  place-items: center;
+  border-radius: 0.375rem;
+  grid-template-columns: repeat(4, minmax(0, 1fr));
+
+}
+
+.user-page .user-info .user-stats div {
+  padding: 0.5rem;
+  border-radius: 0.375rem;
+  width: 90%;
+  background-color: var(--color-2);
+}
+
+.user-page .user-info .user-stats h1 {
+  color: var(--color-important);
+  font-weight: 700;
+  font-size: 0.9rem;
+  margin-bottom: 0 !important;
+  font-family: 'DM Sans', sans-serif;
+}
+
+.user-page .user-info .user-stats h2 {
+  color: var(--font-2);
+  font-weight: 700;
+  font-size: 0.875rem;
+  line-height: 1.25rem;
+  font-family: 'DM Sans', sans-serif;
+}
+
+.user-page .anime-content {
+  display: flex;
+  padding: 1rem;
+  margin-top: 2rem;
+  padding-top: 1rem;
+  border-radius: 0.5rem;
+  flex-direction: column;
+  background-color: var(--color-0);
+}
+
+.user-page .filters-content ul {
+  gap: 5px;
+  padding-left: 10px;
+  display: flex;
+  flex-direction: row;
+}
+
+.user-page .filters-content ul li {
+  gap: 0.5rem;
+  display: flex;
+  cursor: pointer;
+  padding: 0.8rem;
+  padding-bottom: 0.5rem;
+  width: 120px;
+  transition-duration: var(--transition);
+  border-bottom: 1px solid var(--color-2);
+}
+
+.user-page .filters-content ul li.true {
+  border-bottom: 1px solid var(--color-4);
+}
+
+.user-page .filters-content ul li.true h1 {
+  color: var(--color-important);
+}
+
+.user-page .filters-content ul li h1 {
+  margin-top: 0px;
+  margin-bottom: 0px;
+  font-size: 1rem;
+  line-height: 1rem;
+  transition-property: all;
+  transition-duration: var(--transition);
+}
+
+.user-page .filters-content ul li div {
+  opacity: 0;
+  margin-top: 0px;
+  color: var(--font-2);
+  visibility: hidden;
+  transition-property: all;
+  transition-duration: 200ms;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+.user-page .filters-content ul li:hover div {
+  opacity: 1;
+  color: var(--font-2);
+  visibility: visible;
 }
