+++ src/types/anilistAPITypes.tsx
@@ -1,26 +1,26 @@
-import { Media, PageInfo } from "./anilistGraphQLTypes"
+import { Media, PageInfo } from './anilistGraphQLTypes';
 
 export type UserInfo = {
-  id: number
-  name: string
+  id: number;
+  name: string;
   avatar?: {
-    large?: string
-    medium?: string
-  }
-}
+    large?: string;
+    medium?: string;
+  };
+};
 
 export type AnimeData = {
-  media?: Media[]
-  pageInfo?: PageInfo
-}
+  media?: Media[];
+  pageInfo?: PageInfo;
+};
 
 export type ListAnimeData = {
-  id: number | null
-  mediaId: number | null
-  progress?: number | null
-  media: Media
-}
+  id: number | null;
+  mediaId: number | null;
+  progress?: number | null;
+  media: Media;
+};
 
-export type TrendingAnime = AnimeData
-export type MostPopularAnime = AnimeData
-export type CurrentListAnime = ListAnimeData[]
+export type TrendingAnime = AnimeData;
+export type MostPopularAnime = AnimeData;
+export type CurrentListAnime = ListAnimeData[];
